+++ sdk/communication/Azure.Communication.JobRouter/src/Models/RouterJob.cs
@@ -8,22 +8,21 @@ using System.Linq;
 using System.Text.Json;
 using Azure.Core;
 
-namespace Azure.Communication.JobRouter.Models
+namespace Azure.Communication.JobRouter
 {
     [CodeGenModel("RouterJob")]
-    [CodeGenSuppress("RouterJob")]
-    public partial class RouterJob
+    public partial class RouterJob : IUtf8JsonSerializable
     {
-        /// <summary> Initializes a new instance of RouterJob. </summary>
+        /*/// <summary> Initializes a new instance of RouterJob. </summary>
         internal RouterJob()
         {
             AttachedWorkerSelectors = new ChangeTrackingList<RouterWorkerSelector>();
             Assignments = new ChangeTrackingDictionary<string, RouterJobAssignment>();
             _requestedWorkerSelectors = new ChangeTrackingList<RouterWorkerSelector>();
-            _labels = new ChangeTrackingDictionary<string, object>();
-            _tags = new ChangeTrackingDictionary<string, object>();
+            _labels = new ChangeTrackingDictionary<string, BinaryData>();
+            _tags = new ChangeTrackingDictionary<string, BinaryData>();
             _notes = new ChangeTrackingDictionary<string, string>();
-        }
+        }*/
 
         /// <summary>
         /// A set of key/value pairs that are identifying attributes used by the rules engines to make decisions.
@@ -61,13 +60,13 @@ namespace Azure.Communication.JobRouter.Models
         public JobMatchingMode MatchingMode { get; internal set; }
 
         [CodeGenMember("Labels")]
-        internal IDictionary<string, object> _labels
+        internal IDictionary<string, BinaryData> _labels
         {
             get
             {
                 return Labels != null && Labels.Count != 0
-                    ? Labels?.ToDictionary(x => x.Key, x => x.Value?.Value)
-                    : new ChangeTrackingDictionary<string, object>();
+                    ? Labels?.ToDictionary(x => x.Key, x => BinaryData.FromObjectAsJson(x.Value?.Value))
+                    : new ChangeTrackingDictionary<string, BinaryData>();
             }
             set
             {
@@ -75,20 +74,20 @@ namespace Azure.Communication.JobRouter.Models
                 {
                     foreach (var label in value)
                     {
-                        Labels[label.Key] = new LabelValue(label.Value);
+                        Labels[label.Key] = new LabelValue(label.Value.ToObjectFromJson());
                     }
                 }
             }
         }
 
         [CodeGenMember("Tags")]
-        internal IDictionary<string, object> _tags
+        internal IDictionary<string, BinaryData> _tags
         {
             get
             {
                 return Tags != null && Tags.Count != 0
-                    ? Tags?.ToDictionary(x => x.Key, x => x.Value?.Value)
-                    : new ChangeTrackingDictionary<string, object>();
+                    ? Tags?.ToDictionary(x => x.Key, x => BinaryData.FromObjectAsJson(x.Value?.Value))
+                    : new ChangeTrackingDictionary<string, BinaryData>();
             }
             set
             {
@@ -96,7 +95,7 @@ namespace Azure.Communication.JobRouter.Models
                 {
                     foreach (var tag in value)
                     {
-                        Tags[tag.Key] = new LabelValue(tag.Value);
+                        Tags[tag.Key] = new LabelValue(tag.Value.ToObjectFromJson());
                     }
                 }
             }
@@ -138,6 +137,107 @@ namespace Azure.Communication.JobRouter.Models
             {
                 RequestedWorkerSelectors.AddRange(value);
             }
+        }
+
+        void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
+        {
+            writer.WriteStartObject();
+            if (Optional.IsDefined(ChannelReference))
+            {
+                writer.WritePropertyName("channelReference"u8);
+                writer.WriteStringValue(ChannelReference);
+            }
+            if (Optional.IsDefined(ChannelId))
+            {
+                writer.WritePropertyName("channelId"u8);
+                writer.WriteStringValue(ChannelId);
+            }
+            if (Optional.IsDefined(ClassificationPolicyId))
+            {
+                writer.WritePropertyName("classificationPolicyId"u8);
+                writer.WriteStringValue(ClassificationPolicyId);
+            }
+            if (Optional.IsDefined(QueueId))
+            {
+                writer.WritePropertyName("queueId"u8);
+                writer.WriteStringValue(QueueId);
+            }
+            if (Optional.IsDefined(Priority))
+            {
+                writer.WritePropertyName("priority"u8);
+                writer.WriteNumberValue(Priority.Value);
+            }
+            if (Optional.IsDefined(DispositionCode))
+            {
+                writer.WritePropertyName("dispositionCode"u8);
+                writer.WriteStringValue(DispositionCode);
+            }
+            if (Optional.IsCollectionDefined(_requestedWorkerSelectors))
+            {
+                writer.WritePropertyName("requestedWorkerSelectors"u8);
+                writer.WriteStartArray();
+                foreach (var item in _requestedWorkerSelectors)
+                {
+                    writer.WriteObjectValue(item);
+                }
+                writer.WriteEndArray();
+            }
+            if (Optional.IsCollectionDefined(_labels))
+            {
+                writer.WritePropertyName("labels"u8);
+                writer.WriteStartObject();
+                foreach (var item in _labels)
+                {
+                    writer.WritePropertyName(item.Key);
+                    if (item.Value == null)
+                    {
+                        writer.WriteNullValue();
+                        continue;
+                    }
+                    writer.WriteObjectValue(item.Value.ToObjectFromJson());
+                }
+                writer.WriteEndObject();
+            }
+            if (Optional.IsCollectionDefined(_tags))
+            {
+                writer.WritePropertyName("tags"u8);
+                writer.WriteStartObject();
+                foreach (var item in _tags)
+                {
+                    writer.WritePropertyName(item.Key);
+                    if (item.Value == null)
+                    {
+                        writer.WriteNullValue();
+                        continue;
+                    }
+                    writer.WriteObjectValue(item.Value.ToObjectFromJson());
+                }
+                writer.WriteEndObject();
+            }
+            if (Optional.IsCollectionDefined(_notes))
+            {
+                writer.WritePropertyName("notes"u8);
+                writer.WriteStartObject();
+                foreach (var item in _notes)
+                {
+                    writer.WritePropertyName(item.Key);
+                    writer.WriteStringValue(item.Value);
+                }
+                writer.WriteEndObject();
+            }
+            if (Optional.IsDefined(MatchingMode))
+            {
+                writer.WritePropertyName("matchingMode"u8);
+                writer.WriteObjectValue(MatchingMode);
+            }
+            writer.WriteEndObject();
+        }
+
+        internal virtual RequestContent ToRequestContent()
+        {
+            var content = new Utf8JsonRequestContent();
+            content.JsonWriter.WriteObjectValue(this);
+            return content;
         }
     }
 }