+++ sdk/storage/Azure.Storage.DataMovement.Files.Shares/src/ShareFileDestinationCheckpointData.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.IO;
+
+namespace Azure.Storage.DataMovement.Files.Shares
+{
+    internal class ShareFileDestinationCheckpointData : StorageResourceCheckpointData
+    {
+        public override int Length => 0;
+
+        public override void Serialize(Stream stream)
+        {
+        }
+    }
+}