+++ sdk/storage/Azure.Storage.DataMovement/src/LocalDirectoryStorageResourceContainer.cs
@@ -38,6 +38,17 @@ namespace Azure.Storage.DataMovement
         }
 
         /// <summary>
+        /// Internal Constructor for uri
+        /// </summary>
+        /// <param name="uri"></param>
+        internal LocalDirectoryStorageResourceContainer(Uri uri)
+        {
+            Argument.AssertNotNull(uri, nameof(uri));
+            Argument.AssertNotNullOrWhiteSpace(uri.AbsoluteUri, nameof(uri));
+            _uri = uri;
+        }
+
+        /// <summary>
         /// Gets the storage Resource
         /// </summary>
         /// <param name="childPath"></param>
@@ -67,6 +78,16 @@ namespace Azure.Storage.DataMovement
                     yield return new LocalFileStorageResource(fileSystemInfo.FullName);
                 }
             }
+        }
+
+        public override StorageResourceCheckpointData GetSourceCheckpointData()
+        {
+            return new LocalSourceCheckpointData();
+        }
+
+        public override StorageResourceCheckpointData GetDestinationCheckpointData()
+        {
+            return new LocalDestinationCheckpointData();
         }
     }
 }