+++ sdk/storage/Azure.Storage.DataMovement/src/LocalFileStorageResource.cs
@@ -57,11 +57,12 @@ namespace Azure.Storage.DataMovement
         }
 
         /// <summary>
-        /// Intenral Constructor for uri
+        /// Internal Constructor for uri
         /// </summary>
         /// <param name="uri"></param>
         internal LocalFileStorageResource(Uri uri)
         {
+            Argument.AssertNotNull(uri, nameof(uri));
             Argument.AssertNotNullOrWhiteSpace(uri.AbsoluteUri, nameof(uri));
             _uri = uri;
         }
@@ -275,20 +276,12 @@ namespace Azure.Storage.DataMovement
             return Task.FromResult(false);
         }
 
-        /// <summary>
-        /// Gets the source checkpoint data for this resource that will be written to the checkpointer.
-        /// </summary>
-        /// <returns>A <see cref="StorageResourceCheckpointData"/> containing the checkpoint information for this resource.</returns>
-        protected internal override StorageResourceCheckpointData GetSourceCheckpointData()
+        public override StorageResourceCheckpointData GetSourceCheckpointData()
         {
             return new LocalSourceCheckpointData();
         }
 
-        /// <summary>
-        /// Gets the destination checkpoint data for this resource that will be written to the checkpointer.
-        /// </summary>
-        /// <returns>A <see cref="StorageResourceCheckpointData"/> containing the checkpoint information for this resource.</returns>
-        protected internal override StorageResourceCheckpointData GetDestinationCheckpointData()
+        public override StorageResourceCheckpointData GetDestinationCheckpointData()
         {
             return new LocalDestinationCheckpointData();
         }