+++ sdk/storage/Azure.Storage.DataMovement/src/Shared/LocalTransferCheckpointer.cs
@@ -67,6 +67,7 @@ namespace Azure.Storage.DataMovement
                 throw Errors.CollisionTransferIdCheckpointer(transferId);
             }
 
+            bool isContainer = source is StorageResourceContainer;
             JobPlanHeader header = new(
                 DataMovementConstants.JobPlanFile.SchemaVersion,
                 transferId,
@@ -74,10 +75,13 @@ namespace Azure.Storage.DataMovement
                 GetOperationType(source, destination),
                 source.ProviderId,
                 destination.ProviderId,
+                isContainer,
                 false, /* enumerationComplete */
                 new DataTransferStatusInternal(),
-                source.Uri.AbsoluteUri,
-                destination.Uri.AbsoluteUri);
+                source.Uri.ToSanitizedString(),
+                destination.Uri.ToSanitizedString(),
+                source.GetSourceCheckpointData(),
+                destination.GetDestinationCheckpointData());
 
             using (Stream headerStream = new MemoryStream())
             {
@@ -150,13 +154,19 @@ namespace Azure.Storage.DataMovement
             CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
             if (_transferStates.TryGetValue(transferId, out JobPlanFile jobPlanFile))
             {
+                // Lock MMF
+                await jobPlanFile.WriteLock.WaitAsync().ConfigureAwait(false);
+
                 using (MemoryMappedFile mmf = MemoryMappedFile.CreateFromFile(jobPlanFile.FilePath))
                 using (MemoryMappedViewStream mmfStream = mmf.CreateViewStream(offset, length, MemoryMappedFileAccess.Read))
                 {
                     await mmfStream.CopyToAsync(copiedStream).ConfigureAwait(false);
-                    copiedStream.Position = 0;
-                    return copiedStream;
                 }
+
+                // Release MMF
+                jobPlanFile.WriteLock.Release();
+                copiedStream.Position = 0;
+                return copiedStream;
             }
             else
             {