+++ sdk/storage/Azure.Storage.DataMovement/tests/LocalTransferCheckpointerFactory.cs
@@ -68,7 +68,7 @@ namespace Azure.Storage.DataMovement.Tests
                     jobPartCount: _partCountDefault);
             }
 
-            // Return constructed chekcpointer
+            // Return constructed checkpointer
             return new LocalTransferCheckpointer(_checkpointerPath);
         }
 
@@ -152,9 +152,15 @@ namespace Azure.Storage.DataMovement.Tests
             string parentDestinationPath = _testDestinationPath,
             string sourceProviderId = _testSourceProviderId,
             string destinationProviderId = _testDestinationProviderId,
-            DataTransferStatus status = default)
+            bool isContainer = false,
+            DataTransferStatus status = default,
+            StorageResourceCheckpointData sourceCheckpointData = default,
+            StorageResourceCheckpointData destinationCheckpointData = default)
         {
             status ??= new DataTransferStatus();
+            sourceCheckpointData ??= MockResourceCheckpointData.DefaultInstance;
+            destinationCheckpointData ??= MockResourceCheckpointData.DefaultInstance;
+
             JobPlanHeader header = new JobPlanHeader(
                 DataMovementConstants.JobPlanFile.SchemaVersion,
                 transferId,
@@ -162,10 +168,13 @@ namespace Azure.Storage.DataMovement.Tests
                 JobPlanOperation.ServiceToService,
                 sourceProviderId,
                 destinationProviderId,
+                isContainer,
                 false, /* enumerationComplete */
                 status,
                 parentSourcePath,
-                parentDestinationPath);
+                parentDestinationPath,
+                sourceCheckpointData,
+                destinationCheckpointData);
 
             string filePath = Path.Combine(checkpointPath, $"{transferId}.{DataMovementConstants.JobPlanFile.FileExtension}");
             using (FileStream stream = File.Create(filePath))