+++ sdk/eventgrid/Azure.Messaging.EventGrid/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 4.21.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 4.20.0 (2023-10-18)
+
+### Bugs Fixed
+
+- Added several communication models into the correct namespace and hid the old ones.
+
 ## 4.19.0 (2023-10-11)
 
 ### Features Added